MDL-57404 admin_settings: Allow comments in ip block
This adds the ability to to add comments to ip lists in moodle.

Currently, the configiplist textentry box does not allow the user to annotate
the IP addresses defined.  For example, I'd like to be able to define:
 192.168.1.1     # London office
 10.1.1.1        # New york office
 118.209.246.240 # My home IP

This would allow me to revisit this list after a few months, and remove
IP addresses that are no longer required - without having to manually confirm
each IP address